Refactored Profiler to utils package
